health: #184 clean up playground tests a little bit
