--- packages/core/src/test/playgrounds/playground.test.ts
@@ -1,253 +1,0 @@
-import { ExpansionValue } from "@entity-space/common";
-import { Unbox } from "@entity-space/utils";
-import { Instance } from "../../lib/entity/blueprint/instance";
-import { Query } from "../../lib/query/query";
-import { TreeNodeRepository } from "../facade/data";
-import { CanvasModel, UserModel } from "../facade/model";
-
-// const inNumberSet = new InNumberSetCriterion([1, 2, 3]);
-// const inNumberRange = new InStringRangeCriterion(["1", "3"]);
-
-// const reduced = inNumberSet.reduce(inNumberRange);
-// const reducedSet = inNumberSet.reduce(inNumberSet);
-// const reducedRange = inNumberRange.reduce(inNumberRange);
-// type Unbox<T> = T extends Array<infer U> ? U : T;
-// type Box<T, U = any[]> = any[] extends U ? T[] : T;
-
-// credit to captain-yossarian https://captain-yossarian.medium.com/typescript-object-oriented-typings-4fd42ce14c75
-// function Mixin<T extends ClassType, R extends T[]>(...classRefs: [...R]): new (...args: any[]) => UnionToIntersection<InstanceType<[...R][number]>> {
-//     return merge(class {}, ...classRefs);
-// }
-
-describe("prototyping-playground", () => {
-    xit("more expansion stuff", () => {
-        type Expand<T, E> = T extends number | string | null
-            ? Exclude<T, undefined>
-            : T extends any[]
-            ? Expand<T[number], E>[]
-            : "valueOf" extends keyof E // dirty solution, but cleaner for intellisense
-            ? T
-            : T & { [K in keyof (T | E)]-?: Expand<T[K], E[K]> };
-
-        type ExpansionValue<T, U = Unbox<T>> = true | { [K in keyof U]?: ExpansionValue<U[K]> };
-
-        interface Foo {
-            id: number;
-            name: string;
-            bar?: Bar;
-        }
-
-        interface Bar {
-            id: number;
-            name: string;
-            baz?: Baz[];
-        }
-
-        interface BazA {
-            id: number;
-            name: string;
-            type: "baz-a";
-            onlyInA: boolean;
-        }
-
-        interface BazB {
-            id: number;
-            name: string;
-            type: "baz-b";
-            onlyInB: boolean;
-        }
-
-        type Baz = BazA | BazB;
-
-        // function expand<T, E extends ExpansionValue_v2<T>>(expansion: E): Expand<T, E> {
-        function isExpanded<T, E extends ExpansionValue<T>>(entity: T, expansion: E): entity is T & Expand<T, E> {
-            return true;
-        }
-
-        const foo: Foo = {} as any;
-
-        if (isExpanded(foo, { bar: { baz: { onlyInA: true } } })) {
-            const item = foo.bar.baz[0];
-        }
-    });
-    function foo(foo: readonly number[]) {}
-
-    // [todo] investigate this idea - for each class that just wraps data (such as QueryResult), have a Struct/Json interface
-    // that is implemented by the class
-    interface QueryResult_Struct_or_Json_decide_later {
-        readonly queries: readonly Query[];
-    }
-
-    const treeNodeRepo = new TreeNodeRepository();
-    it("generic expand w/ union types", () => {
-        // type Expansion<T> = T extends number | string
-        //     ? true
-        //     : T extends any[]
-        //     ? Expansion<T[number]> | true
-        //     : T extends Record<string, infer U>
-        //     ? Expansion<U>
-        //     : { [K in keyof T]?: Expansion<T[K]> } | true;
-
-        // type Expand<T, E> = T extends number | string
-        //     ? Exclude<T, undefined>
-        //     : T extends any[]
-        //     ? Expand<T[number], E>[]
-        //     : "valueOf" extends keyof E // dirty solution, but cleaner for intellisense
-        //     ? T
-        //     : T extends Record<string, infer X>
-        //     ? Record<string, Expand<X, E>>
-        //     : T & { [K in keyof (T | E)]-?: Expand<T[K], E[K]> };
-
-        interface Square {
-            id: number;
-            area: number;
-            length: number;
-            type: "square";
-        }
-
-        interface Circle {
-            id: number;
-            area: number;
-            radius: number;
-            type: "circle";
-        }
-
-        interface Canvas {
-            id: number;
-            name: string;
-            shapes?: (Square | Circle)[];
-        }
-
-        function takesExpansion<E = ExpansionValue<Instance<CanvasModel>>>(expansion: E): typeof expansion {
-            return {} as any;
-        }
-
-        takesExpansion({ author: {} });
-    });
-
-    it("trying to implement generic expand", () => {
-        interface Brand {
-            id: number;
-            name: string;
-        }
-
-        interface Stuff {
-            id: number;
-            moo?: number;
-        }
-
-        interface Product {
-            id: number;
-            name: string;
-            price: number;
-            rating?: number;
-            brand: Brand;
-            reviews?: ProductReview[];
-            _stuff?: {
-                [k: string]: Stuff;
-            };
-        }
-
-        interface ProductReview {
-            id: number;
-            productId?: number;
-            reviewText?: string;
-            product?: Product;
-        }
-
-        type Expansion<T> = T extends number | string
-            ? true
-            : T extends any[]
-            ? Expansion<T[number]> | true
-            : // : T extends Record<string, infer U>
-              // ? Expansion<U>
-              { [K in keyof T]?: Expansion<T[K]> } | true;
-
-        type Expand<T, E> = T extends number | string
-            ? Exclude<T, undefined>
-            : T extends any[]
-            ? Expand<T[number], E>[]
-            : "valueOf" extends keyof E // dirty solution, but cleaner for intellisense
-            ? T
-            : T extends Record<string, infer X>
-            ? Record<string, Expand<X, E>>
-            : T & { [K in keyof (T | E)]-?: Expand<T[K], E[K]> };
-
-        type X = keyof true;
-        type ExpandedValue = Expand<Product["rating"], "asdasd">;
-        type ExpandedArray = Expand<Product["reviews"], "asdasd">;
-        type MappedProduct = { [K in keyof Product]: number };
-        // type MappedProduct = { [K in keyof Product]: K extends "brand" ? Brand : number };
-
-        // function takesExpansion<E extends Expansion<MappedProduct>>(expansion: E): typeof expansion {
-        type ProductOrReview = Product | ProductReview;
-
-        function takesExpansion<E extends Expansion<ProductOrReview>>(expansion: E): typeof expansion {
-            return {} as any;
-        }
-        // [todo] note somewhere that we cant have Expansion of mapped types w/ keyof; so we need to make model metadata
-        // simpler when getting the Instance type of it
-        function takesExpansion_notWorky<E extends Expansion<Record<keyof Product, true>>>(
-            expansion: E
-        ): typeof expansion {
-            return {} as any;
-        }
-
-        const expansion = takesExpansion({ _stuff: { moo: true }, reviews: { reviewText: true } });
-        takesExpansion_notWorky({});
-
-        type ExpandedProduct = Expand<Product, typeof expansion>;
-        const expandedProduct: ExpandedProduct = {
-            id: 1,
-            name: "foo",
-            price: 64,
-            brand: { id: 2, name: "bar" },
-            _stuff: {
-                foo: { id: 3, moo: 4 },
-            },
-            reviews: [{ id: 123, reviewText: "" }],
-        };
-
-        type UserInstance = Instance<UserModel>;
-        const user: UserInstance = { id: 1, name: "foo" };
-        type UserExpansion = Expansion<UserInstance>;
-
-        function takesUserExpansion_old<E extends UserExpansion>(expansion: E): typeof expansion {
-            return expansion;
-        }
-
-        takesUserExpansion_old({});
-    });
-
-    xit("screwing around with criterion templates", () => {
-        // function instantiateTemplate<T extends CriterionTemplate>(template: T): InstancedCriterionTemplate<T> {
-        //     return {} as any;
-        // }
-        // const instanced_or_inNumberRange = instantiateTemplate(new OrCriteriaTemplate([InNumberRangeCriterion]));
-        // instanced_or_inNumberRange.getItems()[0];
-        // const instanced_or_inNumberRange_inNumberSet = instantiateTemplate(
-        //     new OrCriteriaTemplate([
-        //         InNumberRangeCriterion,
-        //         InNumberSetCriterion,
-        //         new AndCriteriaTemplate([InStringRangeCriterion]),
-        //     ])
-        // );
-        // const item_B = instanced_or_inNumberRange_inNumberSet.getItems()[0];
-        // const instanced_or_propertyCriteria = instantiateTemplate(
-        //     new NamedCriteriaTemplate({
-        //         foo: [InNumberRangeCriterion, new OrCriteriaTemplate([InNumberRangeCriterion])],
-        //         bar: [InNumberSetCriterion, InNumberRangeCriterion],
-        //     })
-        // );
-        // const foo = instanced_or_propertyCriteria.getBag().foo;
-        // const bar = instanced_or_propertyCriteria.getBag().bar;
-        // const instanced_deepMix = instantiateTemplate(
-        //     new OrCriteriaTemplate([
-        //         new NamedCriteriaTemplate({
-        //             foo: [InNumberRangeCriterion, new OrCriteriaTemplate([InNumberRangeCriterion])],
-        //         }),
-        //     ])
-        // );
-        // instanced_deepMix.getItems()[0].getBag().foo;
-    });
-});
